+++ examples/test_examples.py
@@ -1,9 +1,9 @@
 import os
 import sys
-from importlib import reload
+from importlib import metadata, reload
 
-import flask_sqlalchemy
 import pytest
+from packaging.version import Version
 
 full_examples = [
     'basic',
@@ -16,6 +16,7 @@ full_examples = [
 ]
 
 examples = full_examples + ['pagination']
+flask_sqlalchemy_version = Version(metadata.version('flask_sqlalchemy')).major
 
 
 @pytest.fixture
@@ -45,9 +46,7 @@ def test_say_hello(client):
     assert data['message'] == 'Hello!'
 
 
-@pytest.mark.skipif(
-    flask_sqlalchemy.__version__ < '3', reason='Need flask_sqlalchemy 3 for get_or_404'
-)
+@pytest.mark.skipif(flask_sqlalchemy_version < 3, reason='Need flask_sqlalchemy 3 for get_or_404')
 @pytest.mark.parametrize('client', examples, indirect=True)
 def test_get_pet(client):
     rv = client.get('/pets/1')
@@ -115,9 +114,7 @@ def test_create_pet_with_bad_data(client, data):
     assert rv.json
 
 
-@pytest.mark.skipif(
-    flask_sqlalchemy.__version__ < '3', reason='Need flask_sqlalchemy 3 for get_or_404'
-)
+@pytest.mark.skipif(flask_sqlalchemy_version < 3, reason='Need flask_sqlalchemy 3 for get_or_404')
 @pytest.mark.parametrize('client', full_examples, indirect=True)
 def test_update_pet(client):
     new_data = {'name': 'Ghost', 'category': 'dog'}
@@ -150,9 +147,7 @@ def test_update_pet_with_bad_data(client, data):
     assert rv.json
 
 
-@pytest.mark.skipif(
-    flask_sqlalchemy.__version__ < '3', reason='Need flask_sqlalchemy 3 for get_or_404'
-)
+@pytest.mark.skipif(flask_sqlalchemy_version < 3, reason='Need flask_sqlalchemy 3 for get_or_404')
 @pytest.mark.parametrize('client', full_examples, indirect=True)
 def test_delete_pet(client):
     rv = client.delete('/pets/1')
@@ -167,9 +162,7 @@ def test_delete_pet(client):
     assert rv.json
 
 
-@pytest.mark.skipif(
-    flask_sqlalchemy.__version__ < '3', reason='Need flask_sqlalchemy 3 for paginate'
-)
+@pytest.mark.skipif(flask_sqlalchemy_version < 3, reason='Need flask_sqlalchemy 3 for paginate')
 @pytest.mark.parametrize('client', ['pagination'], indirect=True)
 def test_get_pets_pagination(client):
     rv = client.get('/pets')
@@ -199,9 +192,7 @@ def test_get_pets_pagination(client):
     assert rv.json['pets'][-1]['id'] == 20
 
 
-@pytest.mark.skipif(
-    flask_sqlalchemy.__version__ < '3', reason='Need flask_sqlalchemy 3 for paginate'
-)
+@pytest.mark.skipif(flask_sqlalchemy_version < 3, reason='Need flask_sqlalchemy 3 for paginate')
 @pytest.mark.parametrize('client', ['pagination'], indirect=True)
 def test_get_pets_pagination_with_bad_data(client):
     rv = client.get('/pets?per_page=100')